backend/requirements.txt
Django==2.2.19
psycopg2-binary==2.8.6
python-dotenv==0.20.0
djangorestframework==3.12.4
django-filter==2.4.0
djangorestframework-simplejwt==4.8.0
Pillow==9.0.1
djoser==2.1.0
drf-base64==2.0
